+++ taskmonitor/core/celery_queues.py
@@ -1,98 +1,29 @@
 """API for working with celery task queue."""
 
 import concurrent.futures
-import datetime as dt
 import functools
 import itertools
 import json
 from collections import defaultdict
-from dataclasses import dataclass
-from typing import List, NamedTuple, Optional
+from typing import List
 
 import redis
 
 from django.conf import settings
-from django.core.cache import cache
-from django.utils import timezone
 
 from allianceauth.services.hooks import get_extension_logger
 from app_utils.logging import LoggerAddTag
 
 from taskmonitor import __title__
 from taskmonitor.app_settings import TASKMONITOR_QUEUED_TASKS_CACHE_TIMEOUT
-from taskmonitor.helpers import extract_app_name
+
+from .tasks_cache import CacheApi, QueuedTaskShort
 
 PRIORITY_SEP = "\x06\x16"
 DEFAULT_PRIORITY_STEPS = range(10)
 QUEUED_TASKS_CACHE_KEY = "queued-tasks-cache-key"
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)
-
-
-class QueuedTaskShort(NamedTuple):
-    """DTO for queued tasks, optimized for size."""
-
-    app_name: str
-    id: int
-    name: str
-    priority: int
-
-    @classmethod
-    def from_dict(cls, obj: dict) -> "QueuedTaskShort":
-        """Create QueuedTaskShort from raw task dict."""
-        if "headers" not in obj:
-            raise ValueError("headers missing in obj")
-        headers = obj["headers"]
-        properties = obj["properties"] if "properties" in obj else {}
-        task_name = headers["task"]
-        return cls(
-            app_name=extract_app_name(task_name),
-            id=headers["id"],
-            name=task_name,
-            priority=properties.get("priority"),
-        )
-
-
-class QueuedTaskCacheEntry(NamedTuple):
-    """Wrapper for storing queued tasks in cache with additional data."""
-
-    tasks: List[QueuedTaskShort]
-    created_at: dt.datetime
-
-
-@dataclass(frozen=True)
-class CacheApi:
-    """API for working with the cache for queued tasks."""
-
-    cache_key: str
-
-    def get(self) -> Optional[QueuedTaskCacheEntry]:
-        """Retrieve content of cache or None if cache is expired or invalid."""
-        data = cache.get(self.cache_key)
-        if data is not None and not isinstance(data, QueuedTaskCacheEntry):
-            raise TypeError("Expected data to be of type 'QueuedTaskCacheEntry'")
-        return data
-
-    def set(self, tasks: List[QueuedTaskShort]) -> QueuedTaskCacheEntry:
-        """Store given tasks in cache."""
-        data = QueuedTaskCacheEntry(tasks=tasks, created_at=timezone.now())
-        cache.set(
-            self.cache_key,
-            data,
-            timeout=TASKMONITOR_QUEUED_TASKS_CACHE_TIMEOUT,
-        )
-        return data
-
-    def clear(self):
-        """Clear the cache."""
-        cache.delete(self.cache_key)
-
-    def created_at(self) -> Optional[dt.datetime]:
-        """Return date when cache was created or None when cache is invalid."""
-        data = self.get()
-        if data is None:
-            return None
-        return data.created_at
 
 
 def _redis_client():
@@ -186,4 +117,6 @@ def add_tasks(queue_name: str, raw_tasks: list):
     del tasks_by_priority
 
 
-local_cache = CacheApi(QUEUED_TASKS_CACHE_KEY)
+local_cache = CacheApi(
+    cache_key=QUEUED_TASKS_CACHE_KEY, timeout=TASKMONITOR_QUEUED_TASKS_CACHE_TIMEOUT
+)