+++ taskmonitor/tests/core/test_celery_queues.py
@@ -1,10 +1,9 @@
 from unittest import mock
 
-from django.core.cache import cache
 from django.test import TestCase
 
 from taskmonitor.core import celery_queues
-from taskmonitor.core.celery_queues import CacheApi, QueuedTaskShort, local_cache
+from taskmonitor.core.celery_queues import QueuedTaskShort, local_cache
 
 from ..factories import QueuedTaskRawFactory
 
@@ -47,7 +46,7 @@ class TestCeleryQueues(TestCase):
         raw_task_3 = QueuedTaskRawFactory(properties__priority=3)
         celery_queues.add_tasks(CELERY_QUEUE_NAME, [raw_task_1, raw_task_2, raw_task_3])
         # when
-        with mock.patch(MODULE_PATH + ".cache") as m:
+        with mock.patch(MODULE_PATH + ".local_cache") as m:
             m.get.return_value = None
             result = celery_queues.fetch_tasks()
         # then
@@ -74,58 +73,3 @@ class TestCeleryQueues(TestCase):
         result = celery_queues.fetch_tasks()
         # then
         self.assertEqual(result[0], QueuedTaskShort.from_dict(tasks[0]))
-
-
-@mock.patch(MODULE_PATH + ".TASKMONITOR_QUEUED_TASKS_CACHE_TIMEOUT", 60)
-class TestCacheAPi(TestCase):
-    CACHE_KEY = "test-cache-api-cache-key"
-
-    def setUp(self) -> None:
-        cache.delete(self.CACHE_KEY)
-
-    def test_should_store_and_fetch_from_cache(self):
-        # given
-        local_cache = CacheApi(self.CACHE_KEY)
-        tasks = [QueuedTaskShort.from_dict(QueuedTaskRawFactory())]
-        # when
-        local_cache.set(tasks)
-        result = local_cache.get()
-        # then
-        self.assertEqual(result.tasks, tasks)
-
-    def test_should_raise_error_when_cache_returned_wrong_datatype(self):
-        # given
-        local_cache = CacheApi(self.CACHE_KEY)
-        cache.set(key=self.CACHE_KEY, value="abc")
-        # when/then
-        with self.assertRaises(TypeError):
-            local_cache.get()
-
-    def test_should_clear_cache(self):
-        # given
-        local_cache = CacheApi(self.CACHE_KEY)
-        tasks = [QueuedTaskShort.from_dict(QueuedTaskRawFactory())]
-        local_cache.set(tasks)
-        # when
-        local_cache.clear()
-        # then
-        self.assertIsNone(local_cache.get())
-
-    def test_should_return_created_at(self):
-        # given
-        local_cache = CacheApi(self.CACHE_KEY)
-        tasks = [QueuedTaskShort.from_dict(QueuedTaskRawFactory())]
-        local_cache.set(tasks)
-        data = local_cache.get()
-        # when
-        result = local_cache.created_at()
-        # then
-        self.assertEqual(result, data.created_at)
-
-    def test_should_return_none_for_created_at_when_cache_invalid(self):
-        # given
-        local_cache = CacheApi(self.CACHE_KEY)
-        # when
-        result = local_cache.created_at()
-        # then
-        self.assertIsNone(result)